Refactor Firebird's nativeTest
