+++ docs/document/content/user-manual/shardingsphere-jdbc/optional-plugins/clickhouse/_index.cn.md
@@ -43,7 +43,7 @@ ShardingSphere 对 ClickHouse JDBC Driver 的支持位于可选模块中。
 ```yaml
 services:
   clickhouse-server:
-    image: clickhouse/clickhouse-server:25.6.5.41
+    image: clickhouse/clickhouse-server:25.9.6.117
     environment:
       CLICKHOUSE_SKIP_USER_SETUP: "1"
     ports:
