+++ test/native/src/test/java/org/apache/shardingsphere/test/natived/commons/algorithm/testcontainers/type/TcFirebirdDatabaseType.java
@@ -15,32 +15,32 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.test.natived.commons.algorithm.testcontainers;
+package org.apache.shardingsphere.test.natived.commons.algorithm.testcontainers.type;
 
 import org.apache.shardingsphere.database.connector.core.type.DatabaseType;
 import org.apache.shardingsphere.infra.spi.type.typed.TypedSPILoader;
 
+import java.util.Arrays;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.Optional;
 
 /**
- * Database type of PostgreSQL in testcontainers.
+ * Database type of Firebird in testcontainers.
  */
-public final class TcPostgreSQLDatabaseType implements TestcontainersDatabaseType {
+public final class TcFirebirdDatabaseType implements TestcontainersDatabaseType {
     
     @Override
     public Collection<String> getJdbcUrlPrefixes() {
-        return Collections.singleton("jdbc:tc:postgresql:");
+        return Arrays.asList("jdbc:tc:firebird:", "jdbc:tc:firebirdsql:");
     }
     
     @Override
     public Optional<DatabaseType> getTrunkDatabaseType() {
-        return Optional.of(TypedSPILoader.getService(DatabaseType.class, "PostgreSQL"));
+        return Optional.of(TypedSPILoader.getService(DatabaseType.class, "Firebird"));
     }
     
     @Override
     public String getType() {
-        return "TC-PostgreSQL";
+        return "TC-Firebird";
     }
 }
